+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added `NortonGuide.maybe`.
+  ([#11](https://github.com/davep/ngdb.py/pull/11))
+
 ## v0.8.0
 
 **Released: 2025-03-09**